+++ pages/user/[slug].tsx
@@ -16,17 +16,26 @@ export default function Page({ tweets, error, user }: Props) {
    if (error) {
       return (
          <>
-         <Head>
-            <title>Error | tweethub</title>
-            <meta
-               name="viewport"
-               content="width=device-width, initial-scale=1"
-            />
-            <link rel="icon" href="/favicon.ico" />
-         </Head>
-         <div className="flex h-screen items-center justify-center">
-            Can not fetch any data, check your INTERNET or PROXY
-         </div></>
+            <Head>
+               <title>Error | tweethub</title>
+               <meta
+                  name="viewport"
+                  content="width=device-width, initial-scale=1"
+               />
+               <link rel="icon" href="/favicon.ico" />
+            </Head>
+            <div className="flex h-screen items-center justify-center">
+               <div className="flex flex-col items-center justify-center">
+                  <p className="text-center px-2">Can not fetch any data, check your INTERNET or PROXY</p>
+                  <button
+                     className="mt-3 hover:bg-twitter transition hover:text-white rounded-lg border border-twitter bg-transparent px-2 py-1"
+                     onClick={() => window.location.reload()}
+                  >
+                     Reload
+                  </button>
+               </div>
+            </div>
+         </>
       )
    }
 
@@ -37,7 +46,6 @@ export default function Page({ tweets, error, user }: Props) {
    )
 }
 
-
 export const getServerSideProps: GetServerSideProps = async ({ params }) => {
    const userId = params?.slug
    try {
@@ -47,7 +55,7 @@ export const getServerSideProps: GetServerSideProps = async ({ params }) => {
       return {
          props: {
             tweets,
-            user
+            user,
          },
       }
    } catch (e) {
