+++ utils/parser/feedDataParse.ts
@@ -1,11 +1,10 @@
-import { Tweet, Comments } from '@/types/typings'
+import { Tweet, Comments, Likes } from "@/types/typings"
 
 interface TweetData {
    id: number
    attributes: {
       text: string
       blockTweet: boolean
-      likes: number
       createdAt: string
       updatedAt: string
       image?: {
@@ -39,7 +38,6 @@ interface TweetData {
             attributes: {
                comment: string
                blockComment: boolean
-               likes: number
                createdAt: string
                updatedAt: string
                user: {
@@ -62,29 +60,85 @@ interface TweetData {
             }
          }[]
       }
+      likes: {
+         data: {
+            id: number
+            attributes: {
+               username: string
+            }
+         }[]
+      }
    }
 }
 
 export function parseTweetData(tweetData: TweetData[]): Tweet[] {
    return tweetData.map((data) => {
       const { id, attributes } = data
-      const { text, blockTweet, likes, createdAt, updatedAt } = attributes
-      const image = attributes?.image?.data ? attributes.image.data.attributes.url : ""
+      const { text, blockTweet, createdAt, updatedAt } = attributes
+
+      const image = attributes?.image?.data
+         ? attributes.image.data.attributes.url
+         : ""
 
       const { username, blocked } = attributes.user.data.attributes
-      const profileImage = attributes?.user?.data?.attributes?.profileImage?.data?.attributes?.url ? attributes.user.data.attributes.profileImage.data.attributes.url : null
-      const user = { id: attributes.user.data.id, username, blocked, profileImage }
 
-      const comments: Comments[] = attributes.comments.data.map((commentData) => {
-         const { id, attributes } = commentData
-         const { comment, blockComment, likes, createdAt, updatedAt } = attributes
+      const profileImage = attributes?.user?.data?.attributes?.profileImage
+         ?.data?.attributes?.url
+         ? attributes.user.data.attributes.profileImage.data.attributes.url
+         : null
 
-         const { username, blocked } = attributes.user.data.attributes
-         const profileImage = attributes?.user?.data?.attributes?.profileImage?.data?.attributes?.url ? attributes.user.data.attributes.profileImage.data.attributes.url : null
-         const user = { id: attributes.user.data.id, username, blocked, profileImage }
-         return { id, comment, blockComment, likes, createdAt, updatedAt, user }
+      const user = {
+         id: attributes.user.data.id,
+         username,
+         blocked,
+         profileImage,
+      }
+
+      const likes: Likes[] = attributes.likes?.data.map((likeData) => {
+         const username = likeData.attributes.username
+         return {
+            username,
+         }
       })
 
-      return { id, text, blockTweet, likes, createdAt, updatedAt, image, user, comments }
+      const comments: Comments[] = attributes.comments.data.map(
+         (commentData) => {
+            const { id, attributes } = commentData
+            const { comment, blockComment, createdAt, updatedAt } = attributes
+
+            const { username, blocked } = attributes.user.data.attributes
+            const profileImage = attributes?.user?.data?.attributes
+               ?.profileImage?.data?.attributes?.url
+               ? attributes.user.data.attributes.profileImage.data.attributes
+                    .url
+               : null
+            const user = {
+               id: attributes.user.data.id,
+               username,
+               blocked,
+               profileImage,
+            }
+            return {
+               id,
+               comment,
+               blockComment,
+               createdAt,
+               updatedAt,
+               user,
+            }
+         }
+      )
+
+      return {
+         id,
+         text,
+         blockTweet,
+         createdAt,
+         updatedAt,
+         image,
+         user,
+         comments,
+         likes,
+      }
    })
 }
